Make use of new CardLoader context
Change-type: patch
