+++ apps/ui/lib/components/HomeChannel/tests/HomeChannel.spec.jsx
@@ -24,7 +24,6 @@ const sandbox = sinon.createSandbox()
 ava.beforeEach(async () => {
 	context.actions = _.reduce([
 		'addChannel',
-		'getCard',
 		'logout',
 		'queryAPI',
 		'removeView',
