+++ apps/ui/lib/lens/list/SupportThreads.jsx
@@ -286,8 +286,6 @@ export class SupportThreads extends React.Component {
 										return (
 											<CardChatSummary
 												displayOwner
-												selectCard={selectors.getCard}
-												getCard={this.props.actions.getCard}
 												getActor={this.props.actions.getActor}
 												key={card.id}
 												active={
@@ -329,7 +327,6 @@ const mapDispatchToProps = (dispatch) => {
 		actions: redux.bindActionCreators(
 			_.pick(actionCreators, [
 				'getActor',
-				'getCard',
 				'setLensState'
 			]),
 			dispatch
