+++ apps/ui/lib/lens/list/SupportThreadsToAudit/index.js
@@ -27,8 +27,7 @@ const mapDispatchToProps = (dispatch) => {
 	return {
 		actions: redux.bindActionCreators(
 			_.pick(actionCreators, [
-				'getActor',
-				'getCard'
+				'getActor'
 			]),
 			dispatch
 		)
