+++ apps/ui/test/ui-setup.jsx
@@ -21,6 +21,7 @@ import {
 import {
 	configure
 } from 'enzyme'
+import sinon from 'sinon'
 import {
 	MemoryRouter
 } from 'react-router-dom'
@@ -29,6 +30,13 @@ import {
 	CacheProvider
 } from '@emotion/react'
 import createCache from '@emotion/cache'
+
+// Note: importing CardLoaderContext from the root of
+// jellyfish-ui-components results in errors for the 'emotion'
+// package. So instead we import directly from the CardLoader file.
+import {
+	CardLoaderContext
+} from '@balena/jellyfish-ui-components/lib/CardLoader'
 
 import Adapter from 'enzyme-adapter-react-16'
 
@@ -81,7 +89,10 @@ const emotionCache = createCache({
 	key: 'test'
 })
 
-export const getWrapper = (initialState = {}) => {
+export const getWrapper = (initialState = {}, cardLoader = {
+	getCard: sinon.stub().returns(null),
+	selectCard: sinon.stub().returns(sinon.stub().returns(null))
+}) => {
 	const store = mockStore(initialState)
 	return {
 		store,
@@ -94,7 +105,9 @@ export const getWrapper = (initialState = {}) => {
 						<ReduxProvider store={store}>
 							<Provider>
 								<DndProvider backend={HTML5Backend}>
-									{children}
+									<CardLoaderContext.Provider value={cardLoader}>
+										{children}
+									</CardLoaderContext.Provider>
 								</DndProvider>
 							</Provider>
 						</ReduxProvider>
